Dialogue system is done (finally)
So we can set up for an object to trigger a dialogue event. -> Do we want responses to this event? Yes/No doesn't matter. -> Can move onto another dialogue event.

And if you so please, you can make the object repeat the same dialogue (which is the most ideal) OR we have the option to make it say smth diff like a short summary of the message.